refactor(tests): split PT test_args_schemas along the contract-loop seam
tests/squad/penetration_tester/test_args_schemas.py (464) carried two
concerns its own docstring already names: the generic contract loop
(parametrised over MEMBER.schemas, calling the shared assertions) and the
agent-specific accept/reject cases the loop cannot express. Split along
that seam, no coverage change (270 tests, none dropped):

  test_args_schemas.py       <- TestPtArgsSchemaContracts (the generic
                                loop; imports MEMBER + the shared
                                assertions only)
  test_args_schema_cases.py  <- TestSchemaAcceptReject (the StrEnum reject
                                tables, JWT shapes, FQDN-vs-URL, per-family
                                required-field rejection; imports the
                                individual _XArgs only)

The import boundary is clean - MEMBER/assertions on one side, the
individual schema classes on the other - so nothing is duplicated across
the two files.
